create `.initialized` file in the `app` directory to avoid reinitialization

+++ containers/tomcat/bitnami-utils-custom.sh
@@ -1,9 +1,15 @@
 # Tomcat Utility functions
 
 initialize_tomcat_webapps_directory() {
-  echo "==> Initializing Tomcat webapps directory..."
-  echo ""
-  cp -a $BITNAMI_APP_DIR/webapps.defaults/* /app/
+  if [ ! -f /app/.initialized ]; then
+    echo "==> Initializing Tomcat webapps directory..."
+    echo ""
+    cp -a $BITNAMI_APP_DIR/webapps.defaults/* /app/
+    touch /app/.initialized
+  else
+    echo "==> Tomcat webapps directory already initialized..."
+    echo ""
+  fi
 }
 
 print_tomcat_password() {
